chore: use go-task instead of devenv for task running
The idea behind devenv is nice, but as long as reloading any changes
takes seconds to minutes, it's terrible to loop into the development
cycle.

It felt like fitting a square peg in a round hole a lot of the time,
e.g. status checks that write to stdout count as failed. I think it's
trying to do too much, between scripts, tasks, and processes. I'll just
use it for package management and direnv hooks for now.